Address architecture review: extract components, fix N+1 query, remove prop drilling
- Extract InlineCommentForm, InlineFeedbackAnnotation, InlineDraftAnnotation
  into diff-annotations.tsx (~470 lines out of changes-tab.tsx)
- Add GET /agents/:id/reviews/feedback-items endpoint, replace N+1 useQueries
  fan-out with single useQuery
- Replace autoExpandReviewId prop drilling with ?expandReview URL search param
- Replace hand-rolled dropdown with shadcn DropdownMenu
- Replace queryClient.getQueriesData cache read with useAgentDiff hook
- Replace raw fetch() with api() helper
- Unify DraftComment type as re-export of PersistedDraftComment
